Move isDrug method to org.gk.model.InstanceUtilities.
